Extract env/process/clock i64 lowering into host_env_proc_clock.rs (#1254)
Sixteenth cranelift_backend.rs split under #1254. The closure-selection rule
seeded on the env, process, clock tokens identified 17 i64-lowering functions:
- lower_i64_env_option_match_stmt + 4 helper fns (prints_binding_verbatim,
  payload_uses_len_only, stmt_uses_payload_len_only, expr_uses_payload_len_only)
- lower_i64_env_option_call_let_stmts
- lower_i64_env_option_match_value_expr + lower_i64_env_some_arm_expr
  + i64_env_len_expr + i64_env_get_key
- lower_i64_clock_intrinsic_expr
- lower_i64_duration_ms_expr + lower_i64_instant_ms_expr (called by clock fn)
- lower_i64_process_intrinsic_expr
- i64_audited_env_expr + i64_audited_process_expr + i64_audited_clock_expr

Per the #1282 recipe: this is a pure function move like host_net_http. Shared
structs (SpikeEnv, I64StaticBindings), helper fns (lower_i64_string_option_len_call_let_stmts),
and i64_known_static_env stay in the parent because callers cross slice boundaries
(eval_expr, host_fs). All moved fns are pub(crate); mod and pub(crate) use re-export
added after existing host_* peers.

Drops cranelift_backend.rs from 20914 to 20338 lines (576 removed). New file:
host_env_proc_clock.rs at 586 lines. Lowered all tracked ceilings in the same PR.

Do not run cargo fmt on these backend files.

Refs #1254.
